+++ 04/sub_02/src/shared_data/shared_data.h
@@ -3,7 +3,7 @@
 
 void shared_data_update_val(long long data);
 
-uint64_t shared_data_get_val(void);
+long long shared_data_get_val(void);
 
 void shared_data_lock(void);
